requirements.txt
pymongo==4.6.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
email-validator==2.1.0.post1 